refactor(plugin-image): Refactor Pixabay Search component to be more self contained
